+++ packages/benchmark-js/src/index.ts
@@ -1,4 +1,3 @@
 export * from './Benchmark';
 export * from './BenchmarkGroup';
-export { BenchmarkLoggerLevel } from './tools/Logger';
 export * from './types';